requirements.txt
streamlit==1.57.0
ultralytics==8.4.43
pillow
numpy
opencv-python-headless
pandas
matplotlib
torch
torchvision